+++ bundles/admin/appsetup/index.js
@@ -1,0 +1,5 @@
+import './instance';
+import './resources/scss/style.scss';
+
+// register create function for bundleid
+Oskari.bundle('appsetup', () => Oskari.clazz.create('Oskari.admin.bundle.appsetup.AppSetupAdminBundleInstance'));